Move ivy.xml generation stuff to lm-coursier
So that it can be re-used from sbt
