+++ modules/sbt-coursier-shared/src/main/scala/coursier/sbtcoursiershared/SbtCoursierShared.scala
@@ -174,7 +174,7 @@ object SbtCoursierShared extends AutoPlugin {
       versionReconciliation := Seq.empty
     ) ++ {
       if (pubSettings)
-        IvyXml.generateIvyXmlSettings()
+        IvyXmlGeneration.generateIvyXmlSettings()
       else
         Nil
     }
